Make sure the return value is no longer of a Algebraic type
This is usually the case for the spectral invariants. This leads to the following error when a division is needed of these elements:

    TypeError: no canonical coercion from Algebraic Field to Rational Field

+++ sage/matrixtheory.py
@@ -18,20 +18,20 @@ def permanent(m):
     return m.permanent()
 
 def maximum_eigenvalue(m):
-    return max(m.eigenvalues())
+    return float(max(m.eigenvalues()))
 
 def minimum_eigenvalue(m):
-    return min(m.eigenvalues())
+    return float(min(m.eigenvalues()))
 
 def average_eigenvalue(m):
     evs = m.eigenvalues()
-    return sum(evs)/len(evs)
+    return float(sum(evs)/len(evs))
 
 def number_of_distinct_eigenvalues(m):
     return len(set(m.eigenvalues()))
 
 def spectral_radius(m):
-    return max(abs(e) for e in m.eigenvalues())
+    return float(max(abs(e) for e in m.eigenvalues()))
 
 def frobenius_norm(m):
     return sqrt(sum(e*e for e in m.list()))
@@ -47,11 +47,11 @@ def max_column_sum(m):
 
 def ratio_min_max_absolute_eigenvalues(m):
     aevs = [abs(ev) for ev in m.eigenvalues()]
-    return max(aevs)/min(aevs)
+    return float(max(aevs)/min(aevs))
 
 def separator(m):
     sevs = sorted(m.eigenvalues())
-    return sevs[-1] - sevs[-2]
+    return float(sevs[-1] - sevs[-2])
 
 
 invariants = [determinant, nullity, rank, trace, nrows, permanent,
